+++ pkg/cluster/sync.go
@@ -560,7 +560,7 @@ func (c *Cluster) syncRoles() (err error) {
 		userNames = append(userNames, u.Name)
 	}
 
-	if c.needMasterConnectionPooler() || c.needReplicaConnectionPooler() {
+	if needMasterConnectionPooler(&c.Spec) || needReplicaConnectionPooler(&c.Spec) {
 		connectionPoolerUser := c.systemUsers[constants.ConnectionPoolerUserKeyName]
 		userNames = append(userNames, connectionPoolerUser.Name)
 
@@ -837,238 +837,4 @@ func (c *Cluster) syncLogicalBackupJob() error {
 	}
 
 	return nil
-}
-
-func (c *Cluster) syncConnectionPooler(oldSpec,
-	newSpec *acidv1.Postgresql,
-	lookup InstallFunction) (SyncReason, error) {
-
-	var reason SyncReason
-	var err error
-	var newNeedConnectionPooler, oldNeedConnectionPooler bool
-
-	// Check and perform the sync requirements for each of the roles.
-	for _, role := range [2]PostgresRole{Master, Replica} {
-		if role == Master {
-			newNeedConnectionPooler = c.needMasterConnectionPoolerWorker(&newSpec.Spec)
-			oldNeedConnectionPooler = c.needMasterConnectionPoolerWorker(&oldSpec.Spec)
-		} else {
-			newNeedConnectionPooler = c.needReplicaConnectionPoolerWorker(&newSpec.Spec)
-			oldNeedConnectionPooler = c.needReplicaConnectionPoolerWorker(&oldSpec.Spec)
-		}
-		if c.ConnectionPooler == nil {
-			c.ConnectionPooler = &ConnectionPoolerObjects{}
-			c.ConnectionPooler.Deployment = make(map[PostgresRole]*appsv1.Deployment)
-			c.ConnectionPooler.Service = make(map[PostgresRole]*v1.Service)
-		}
-
-		if newNeedConnectionPooler {
-			// Try to sync in any case. If we didn't needed connection pooler before,
-			// it means we want to create it. If it was already present, still sync
-			// since it could happen that there is no difference in specs, and all
-			// the resources are remembered, but the deployment was manually deleted
-			// in between
-			c.logger.Debug("syncing connection pooler")
-
-			// in this case also do not forget to install lookup function as for
-			// creating cluster
-			if !oldNeedConnectionPooler || !c.ConnectionPooler.LookupFunction {
-				newConnectionPooler := newSpec.Spec.ConnectionPooler
-
-				specSchema := ""
-				specUser := ""
-
-				if newConnectionPooler != nil {
-					specSchema = newConnectionPooler.Schema
-					specUser = newConnectionPooler.User
-				}
-
-				schema := util.Coalesce(
-					specSchema,
-					c.OpConfig.ConnectionPooler.Schema)
-
-				user := util.Coalesce(
-					specUser,
-					c.OpConfig.ConnectionPooler.User)
-
-				if err = lookup(schema, user); err != nil {
-					return NoSync, err
-				}
-			}
-
-			if reason, err = c.syncConnectionPoolerWorker(oldSpec, newSpec, role); err != nil {
-				c.logger.Errorf("could not sync connection pooler: %v", err)
-				return reason, err
-			}
-		} else {
-			// Lookup function installation seems to be a fragile point, so
-			// let's log for debugging if we skip it
-			msg := "Skip lookup function installation, old: %d, already installed %d"
-			c.logger.Debug(msg, oldNeedConnectionPooler, c.ConnectionPooler.LookupFunction)
-		}
-
-		if oldNeedConnectionPooler && !newNeedConnectionPooler {
-			// delete and cleanup resources
-			otherRole := role
-			if len(c.RolesConnectionPooler()) == 2 {
-				if role == Master {
-					otherRole = Replica
-				} else {
-					otherRole = Master
-				}
-			}
-			if c.ConnectionPooler != nil &&
-				(c.ConnectionPooler.Deployment[role] != nil ||
-					c.ConnectionPooler.Service[role] != nil) {
-
-				if err = c.deleteConnectionPooler(role); err != nil {
-					c.logger.Warningf("could not remove connection pooler: %v", err)
-				}
-			}
-			if c.ConnectionPooler != nil && c.ConnectionPooler.Deployment[otherRole] == nil && c.ConnectionPooler.Service[otherRole] == nil {
-				c.ConnectionPooler = nil
-			}
-		}
-
-		if !oldNeedConnectionPooler && !newNeedConnectionPooler {
-			// delete and cleanup resources if not empty
-			otherRole := role
-			if len(c.RolesConnectionPooler()) == 2 {
-				if role == Master {
-					otherRole = Replica
-				} else {
-					otherRole = Master
-				}
-			}
-			if c.ConnectionPooler != nil &&
-				(c.ConnectionPooler.Deployment[role] != nil ||
-					c.ConnectionPooler.Service[role] != nil) {
-
-				if err = c.deleteConnectionPooler(role); err != nil {
-					c.logger.Warningf("could not remove connection pooler: %v", err)
-				}
-			} else if c.ConnectionPooler.Deployment[otherRole] == nil && c.ConnectionPooler.Service[otherRole] == nil {
-				c.ConnectionPooler = nil
-			}
-		}
-	}
-
-	return reason, nil
-}
-
-// Synchronize connection pooler resources. Effectively we're interested only in
-// synchronizing the corresponding deployment, but in case of deployment or
-// service is missing, create it. After checking, also remember an object for
-// the future references.
-func (c *Cluster) syncConnectionPoolerWorker(oldSpec, newSpec *acidv1.Postgresql, role PostgresRole) (
-	SyncReason, error) {
-
-	deployment, err := c.KubeClient.
-		Deployments(c.Namespace).
-		Get(context.TODO(), c.connectionPoolerName(role), metav1.GetOptions{})
-
-	if err != nil && k8sutil.ResourceNotFound(err) {
-		msg := "Deployment %s for connection pooler synchronization is not found, create it"
-		c.logger.Warningf(msg, c.connectionPoolerName(role))
-
-		deploymentSpec, err := c.generateConnectionPoolerDeployment(&newSpec.Spec, role)
-		if err != nil {
-			msg = "could not generate deployment for connection pooler: %v"
-			return NoSync, fmt.Errorf(msg, err)
-		}
-
-		deployment, err := c.KubeClient.
-			Deployments(deploymentSpec.Namespace).
-			Create(context.TODO(), deploymentSpec, metav1.CreateOptions{})
-
-		if err != nil {
-			return NoSync, err
-		}
-		c.ConnectionPooler.Deployment[role] = deployment
-	} else if err != nil {
-		msg := "could not get connection pooler deployment to sync: %v"
-		return NoSync, fmt.Errorf(msg, err)
-	} else {
-		c.ConnectionPooler.Deployment[role] = deployment
-
-		// actual synchronization
-		oldConnectionPooler := oldSpec.Spec.ConnectionPooler
-		newConnectionPooler := newSpec.Spec.ConnectionPooler
-
-		// sync implementation below assumes that both old and new specs are
-		// not nil, but it can happen. To avoid any confusion like updating a
-		// deployment because the specification changed from nil to an empty
-		// struct (that was initialized somewhere before) replace any nil with
-		// an empty spec.
-		if oldConnectionPooler == nil {
-			oldConnectionPooler = &acidv1.ConnectionPooler{}
-		}
-
-		if newConnectionPooler == nil {
-			newConnectionPooler = &acidv1.ConnectionPooler{}
-		}
-
-		c.logger.Infof("Old: %+v, New %+v", oldConnectionPooler, newConnectionPooler)
-
-		specSync, specReason := c.needSyncConnectionPoolerSpecs(oldConnectionPooler, newConnectionPooler)
-		defaultsSync, defaultsReason := c.needSyncConnectionPoolerDefaults(newConnectionPooler, deployment)
-		reason := append(specReason, defaultsReason...)
-
-		if specSync || defaultsSync {
-			c.logger.Infof("Update connection pooler deployment %s, reason: %+v",
-				c.connectionPoolerName(role), reason)
-			newDeploymentSpec, err := c.generateConnectionPoolerDeployment(&newSpec.Spec, role)
-			if err != nil {
-				msg := "could not generate deployment for connection pooler: %v"
-				return reason, fmt.Errorf(msg, err)
-			}
-
-			oldDeploymentSpec := c.ConnectionPooler.Deployment[role]
-
-			deployment, err := c.updateConnectionPoolerDeployment(
-				oldDeploymentSpec,
-				newDeploymentSpec,
-				role)
-
-			if err != nil {
-				return reason, err
-			}
-			c.ConnectionPooler.Deployment[role] = deployment
-
-			return reason, nil
-		}
-	}
-
-	newAnnotations := c.AnnotationsToPropagate(c.ConnectionPooler.Deployment[role].Annotations)
-	if newAnnotations != nil {
-		c.updateConnectionPoolerAnnotations(newAnnotations, role)
-	}
-
-	service, err := c.KubeClient.
-		Services(c.Namespace).
-		Get(context.TODO(), c.connectionPoolerName(role), metav1.GetOptions{})
-
-	if err != nil && k8sutil.ResourceNotFound(err) {
-		msg := "Service %s for connection pooler synchronization is not found, create it"
-		c.logger.Warningf(msg, c.connectionPoolerName(role))
-
-		serviceSpec := c.generateConnectionPoolerService(&newSpec.Spec, role)
-		service, err := c.KubeClient.
-			Services(serviceSpec.Namespace).
-			Create(context.TODO(), serviceSpec, metav1.CreateOptions{})
-
-		if err != nil {
-			return NoSync, err
-		}
-		c.ConnectionPooler.Service[role] = service
-
-	} else if err != nil {
-		msg := "could not get connection pooler service to sync: %v"
-		return NoSync, fmt.Errorf(msg, err)
-	} else {
-		// Service updates are not supported and probably not that useful anyway
-		c.ConnectionPooler.Service[role] = service
-	}
-
-	return NoSync, nil
 }
